requirements.txt
numpy>=1.19.0
scipy>=1.5.0
scikit-learn>=0.24.2
pandas>=1.1.5
imbalanced-learn>=0.8.0
joblib>=1.0.0
pyserial>=3.5
matplotlib>=3.3.4
seaborn>=0.11.0
pytest>=6.2.0
adafruit-circuitpython-scd4x>=1.0.0
sparkfun-qwiic-ens160>=1.0.0
